+++ NebulaWorld/MultiplayerSession.cs
@@ -9,6 +9,7 @@ using NebulaWorld.Player;
 using NebulaWorld.Statistics;
 using NebulaWorld.Trash;
 using NebulaWorld.Universe;
+using NebulaWorld.Warning;
 using System;
 
 namespace NebulaWorld
@@ -31,6 +32,7 @@ namespace NebulaWorld
         public TrashManager Trashes { get; private set; }
         public DysonSphereManager DysonSpheres { get; private set; }
         public LaunchManager Launch { get; private set; }
+        public WarningManager Warning { get; private set; }
 
         // Some Patch Flags
         public bool IsTankWindowPointerPress { get; set; }
@@ -69,6 +71,7 @@ namespace NebulaWorld
             Trashes = new TrashManager();
             DysonSpheres = new DysonSphereManager();
             Launch = new LaunchManager();
+            Warning = new WarningManager();
         }
 
         public void Dispose()
@@ -120,6 +123,9 @@ namespace NebulaWorld
 
             Launch?.Dispose();
             Launch = null;
+
+            Warning?.Dispose();
+            Warning = null;
         }
 
         public void OnGameLoadCompleted()
@@ -128,6 +134,11 @@ namespace NebulaWorld
             {
                 Log.Info("Game load completed");
                 IsGameLoaded = true;
+
+                if (Multiplayer.Session.LocalPlayer.IsHost)
+                {
+                    GameMain.history.universeObserveLevel = SimulatedWorld.GetUniverseObserveLevel();
+                }
 
                 if (Multiplayer.Session.LocalPlayer.IsInitialDataReceived)
                 {