+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+0.7.10:
+
+- @starfish: Added WarningSystem syncing
+- @PhantomGamers: Fixed case of NRE when arriving on another planet
+- @PhantomGamers: Fixed issue where Universe Exploration tech would break while in a multiplayer game
+
 0.7.9:
 
 - @sp00ktober: gracefully tell older nebula versions that there is a mod version missmatch.